chore: 가용자금 파트 그래프 삭제
