+++ src/components/screen/LedgerScreen/AvailableFundsOverview/AvailableFundsBottomElement.tsx
@@ -1,33 +1,94 @@
-import AvailableFundsHeader from 'components/screen/LedgerScreen/AvailableFundsOverview/AvailableFundsHeader';
-import AvailableFundsDetail from 'components/screen/LedgerScreen/AvailableFundsOverview/AvailableFundsDetail';
+import {View, StyleSheet} from 'react-native';
+import {theme} from 'styles';
+import {formatToLocaleString} from 'utils/formatAmountValue';
+import Typography from 'components/@common/Typography';
 
 /**
  * @param availableFunds 가용 자금
- * @param cashAndLiquidAssets 현금 및 현금성 자산
+ * @param cash 현금
  * @param savingsAccount 보통 예금
  * @param plannedExpenditure 지출 예정 자금
  */
 export type AvailableFundsBottomElementProps = {
   availableFunds: number;
-  cashAndLiquidAssets: number;
+  cash: number;
   savingsAccount: number;
   plannedExpenditure: number;
 };
 
 export default function AvailableFundsBottomElement({
   availableFunds,
-  cashAndLiquidAssets,
+  cash,
   savingsAccount,
   plannedExpenditure,
 }: AvailableFundsBottomElementProps) {
   return (
-    <>
-      <AvailableFundsHeader availableFunds={availableFunds} />
-      <AvailableFundsDetail
-        cashAndLiquidAssets={cashAndLiquidAssets}
-        savingsAccount={savingsAccount}
-        plannedExpenditure={plannedExpenditure}
-      />
-    </>
+    <View>
+      <View style={availableFundsBottomElementStyles.availableFundsContainer}>
+        <Typography type={'Body2Semibold'} color={'gray6'}>
+          보민님의 순자산
+        </Typography>
+        <Typography type={'Title1Semibold2'}>
+          {formatToLocaleString(availableFunds)}원
+        </Typography>
+      </View>
+      <View style={availableFundsBottomElementStyles.assetDebtContainer}>
+        <View style={availableFundsBottomElementStyles.cashContainer}>
+          <Typography type={'Body1Semibold'} color={'gray6'}>
+            현금
+          </Typography>
+          <Typography type={'Body1Semibold'} color={'green'}>
+            {formatToLocaleString(cash)}원
+          </Typography>
+        </View>
+        <View style={availableFundsBottomElementStyles.savingsAccountContainer}>
+          <Typography type={'Body1Semibold'} color={'gray6'}>
+            보통예금
+          </Typography>
+          <Typography type={'Body1Semibold'} color={'green'}>
+            {formatToLocaleString(savingsAccount)}원
+          </Typography>
+        </View>
+        <View
+          style={availableFundsBottomElementStyles.plannedExpenditureContainer}>
+          <Typography type={'Body1Semibold'} color={'gray6'}>
+            카드대금
+          </Typography>
+          <Typography type={'Body1Semibold'} color={'pink'}>
+            {formatToLocaleString(plannedExpenditure)}원
+          </Typography>
+        </View>
+      </View>
+    </View>
   );
 }
+
+const availableFundsBottomElementStyles = StyleSheet.create({
+  availableFundsContainer: {
+    width: '100%',
+    borderRadius: 18,
+    backgroundColor: theme.palette.gray2,
+    padding: 18,
+    marginBottom: 18,
+  },
+  assetDebtContainer: {
+    display: 'flex',
+    flexDirection: 'column',
+    gap: 10,
+  },
+  cashContainer: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  savingsAccountContainer: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  plannedExpenditureContainer: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+});
